ingestion/requirements.txt
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
camelot-py[cv]==0.11.0
tabula-py==2.8.2
PyPDF2==3.0.1
openai==1.6.1
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.3
